Modificações no código (refatoração, ajuste de serviço, ajuste na pagina de cartão de crédito, inserção de regra de juros)

+++ lib/models/ticket_query/ticket_query.dart
@@ -9,7 +9,7 @@ class TicketQuery {
   String? digitable;
   int? transactionId;
   int? type;
-  int? value;
+  double? value;
   Null? maxValue;
   Null? minValue;
   String? errorCode;
